# tracing: re-use childrenMetadata map across trace spans