@nvanbenschoten 2022-08-16
This closes the gap on heap allocation count on most of these microbenchmarks, leading to a new reduction thanks to other optimizations made during this release cycle. However, the microbenchmarks still show an increase in runtime across the board, which is consistent with Erik's testing.

```
➜ benchdiff --count=10 --old='origin/release-22.1' --post-checkout='dev generate go' --run='KV/././rows=1$$' ./pkg/sql/tests

name                        old time/op    new time/op    delta
KV/Update/SQL/rows=1-10        170µs ± 7%     170µs ± 2%     ~     (p=0.739 n=10+10)
KV/Insert/SQL/rows=1-10        123µs ± 1%     127µs ± 2%   +2.68%  (p=0.000 n=8+9)
KV/Insert/Native/rows=1-10    39.8µs ± 1%    41.6µs ± 2%   +4.49%  (p=0.000 n=10+10)
KV/Delete/Native/rows=1-10    39.5µs ± 1%    41.4µs ± 1%   +4.68%  (p=0.000 n=10+10)
KV/Update/Native/rows=1-10    63.2µs ± 2%    66.3µs ± 2%   +4.88%  (p=0.000 n=10+10)
KV/Scan/SQL/rows=1-10         87.6µs ± 2%    92.8µs ± 2%   +5.93%  (p=0.000 n=10+9)
KV/Scan/Native/rows=1-10      15.7µs ± 2%    17.1µs ± 1%   +9.06%  (p=0.000 n=10+10)
KV/Delete/SQL/rows=1-10        125µs ± 3%     138µs ± 4%  +10.58%  (p=0.000 n=10+9)

name                        old alloc/op   new alloc/op   delta
KV/Insert/Native/rows=1-10    15.8kB ± 0%    15.9kB ± 0%   +0.34%  (p=0.000 n=9+10)
KV/Delete/Native/rows=1-10    15.6kB ± 0%    15.7kB ± 0%   +0.54%  (p=0.000 n=9+10)
KV/Update/Native/rows=1-10    22.3kB ± 0%    22.5kB ± 1%   +0.82%  (p=0.000 n=10+10)
KV/Update/SQL/rows=1-10       51.2kB ± 1%    51.7kB ± 0%   +0.98%  (p=0.000 n=10+10)
KV/Insert/SQL/rows=1-10       44.0kB ± 0%    44.8kB ± 0%   +1.90%  (p=0.000 n=10+10)
KV/Scan/Native/rows=1-10      7.48kB ± 0%    7.65kB ± 0%   +2.28%  (p=0.000 n=9+10)
KV/Scan/SQL/rows=1-10         22.3kB ± 0%    24.3kB ± 0%   +9.27%  (p=0.000 n=9+10)
KV/Delete/SQL/rows=1-10       46.3kB ± 0%    51.8kB ± 0%  +11.85%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
KV/Update/SQL/rows=1-10          534 ± 0%       520 ± 0%   -2.62%  (p=0.000 n=9+8)
KV/Insert/Native/rows=1-10       131 ± 0%       128 ± 0%   -2.29%  (p=0.000 n=10+10)
KV/Update/Native/rows=1-10       186 ± 0%       182 ± 0%   -2.15%  (p=0.000 n=10+8)
KV/Scan/Native/rows=1-10        55.0 ± 0%      54.0 ± 0%   -1.82%  (p=0.000 n=9+10)
KV/Insert/SQL/rows=1-10          365 ± 0%       359 ± 0%   -1.73%  (p=0.000 n=10+10)
KV/Delete/Native/rows=1-10       129 ± 0%       127 ± 0%   -1.55%  (p=0.000 n=9+10)
KV/Scan/SQL/rows=1-10            269 ± 0%       279 ± 0%   +3.72%  (p=0.000 n=10+9)
KV/Delete/SQL/rows=1-10          349 ± 0%       386 ± 0%  +10.60%  (p=0.000 n=10+8)
```